refactor validations
